+++ src/components/hero/campaignHeroSubText.tsx
@@ -6,7 +6,7 @@ import {
 } from "@nypl/design-system-react-components";
 
 const CampaignHeroSubText = ({ featuredItem }: any) => {
-  console.log("featuredItem in CampaignHeroSubText is: ", featuredItem);
+  // console.log("featuredItem in CampaignHeroSubText is: ", featuredItem);
   return (
     <>
       {/* To Do: make the link color blue: */}